requirements.txt
streamlit
numpy
matplotlib
opencv-python
pillow
scipy
scikit-image
